Fix SaveChangesAsync pattern, simplify nobodies.team repo methods, move DTOs from interface file
- Replace ICommunicationPreferenceRepository.SaveChangesAsync with UpdateAsync(preference) to encapsulate EF internals; update CommunicationPreferenceService callers
- Collapse 4 nobodies.team-specific IUserEmailRepository methods into 1 GetAllVerifiedNobodiesTeamEmailsAsync bulk method; move per-user filtering to UserEmailService (in-memory at ~500 users)
- Move ProfileSaveRequest, UserSearchResult, HumanSearchResult out of IProfileService.cs into src/Humans.Application/DTOs/; add missing DTOs using to CachingProfileService and SearchResultMappingExtensions
