+++ src/Humans.Web/Extensions/SearchResultMappingExtensions.cs
@@ -1,3 +1,4 @@
+using Humans.Application.DTOs;
 using Humans.Application.Interfaces;
 using Humans.Web.Controllers;
 using Humans.Web.Models;
